Fix QtQml crashing with GHS compiler when using AOT
In order to support compilers that do not support the packing attribute,
replace the use of the pack macros with static assertions about the
structure size, to ensure that what is generated by host tools
(qmlcachegen, qqc) is compatible with what's loaded by the target at
run-time. This requires padding and re-ordering some structures.

Task-number: QTBUG-61468
Change-Id: I3d82457f086a9b066d1c6df4c46d9f154dd5f208
Reviewed-by: Lars Knoll <lars.knoll@qt.io>