Update the cursor rectangle when select all is triggered
When select all is triggered then the cursor rectangle should be
updated to ensure that the selection handles are visible if they are
displayed on a given platform. For instance, on iOS the loupe handles
should appear in this case.

This also has the added effect that when using Backspace on the iOS
keyboard, it will delete the selected text as opposed to just a
single character as it will handle that situation correctly.

The tst_qquicktextedit::inputMethodUpdate() test is modified to
account for the extra event sent now when selectAll() is called.

Task-number: QTBUG-63835
Change-Id: I94fb0576b286c006dd12c65d0b322d712ed2a96f
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>