Fix crash with loading cache files originating from top-level components
If a .qml file starts with Component {} and its item(s) define their own
properties, alias, etc. then loading this file initially would work, but
loading it from a cache file would crash with dangling pointers in the
property cache. This was due to us registering aliases, properties, etc.
twice in the property cache, exceeding the reservation in the property
cache vectors.

The minimal fix is to skip the root object in the property cache
creating loop as we do handle it separately afterwards. It needs to be
separate because the first object index within the component does not
stem from a binding.

However as the root object index is always zero, I'll make a follow-up
patch to get rid of of the variable.

Task-number: QTBUG-62263
Change-Id: I86b76d38cb490750a561eac2b0ad6fff6ef2e20a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>