Fix argument order for bound signal expressions
Function::updateInternalClass added the arguments in reverse order to
the call context object. However, the index of the member is used as the
index into CallData::args, so the result would be a reverse order when
lookup was done.

Now the order is as expected: first argument first. To handle
duplicates, another pass is done before adding those arguments so that
names are uniques out.

Change-Id: Icfb6af2769b7669ff5589ef29f16c23c314758e8
Reviewed-by: Lars Knoll <lars.knoll@qt.io>