Fix tst_palette
The Default is being ported to use palettes. tst_palette starts
failing as soon as QQuickDefaultTheme::palette() returns a custom
palette, because the expected palette in tst_palette's comparison
was the default platform palette, not the default style palette.
First of all, the expected palette was constructed before any
QtQuick Controls were imported. Secondly, it was not queried from
the platform theme as appropriate.

Change-Id: Ia28872fb2769c6cd5ff0159077769951def17f43
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>