Default: make Dial use palettes
NOTE: Now that we have configurable palettes, we cannot assume that
the default color of the indicator image matches the default palette
color. Therefore we need to make ColorImage aware of the default color
to avoid unnecessary expensive colorizing (see f0697c6). An extra QML
property assignment of a constant value is much cheaper than colorizing
the image.

Starting from f0697c6, the indicator color has been defined like this:

    color: control.visualFocus ? Default.focusColor : undefined

This assumes that the image has a default color of #353637 when not
focused. With this, qmlbench delegates_dial.qml gives a result of 76
frames.

The following options were considered:

A)
  color: visualFocus ? palette.highlight : palette.dark

  => 71 frames

B)
  color: visualFocus ? palette.highlight :
         palette.dark !== "#353637" ? palette.dark : "transparent"

  => 71 frames

C)
  defaultColor: "#353637"
  color: visualFocus ? palette.highlight : palette.dark

  => 75 frames

The last option was a clear winner.

Task-number: QTBUG-63369
Change-Id: Ib8d56724de55b874d1c4ecdf0c09b91cfd2dc3df
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>