skip part of tst_qquickwindow::testWindowVisibilityOrder on Unity
It's flaky: on the CI machines, the first three windows are shown
stacked on top of each other, and none of them has focus, because
the terminal which is running the tests continues to have focus.
If this happened to a user, he would just click in whichever window
he wants to have focus.  (Good thing it's visible, at least.)
So this looks like broken window manager behavior.  Ubuntu will switch
to Gnome soon anyway; then maybe the behavior will be different.

Task-number: QTBUG-62604
Change-Id: Id0db8a61e8335e7f591a38e07c488c055b236239
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>