Default: restore the style singleton attributes
It is known that people have made copies of the .qml files in order to
customize them to their liking. Therefore it is best to restore the old
API for now as it was in 391cba9a, before the palette changes. We will
deprecate and remove it gradually, to give people time to migrate to
palettes.

Task-number: QTBUG-63369
Change-Id: Ib0c2166b00115b1e0ca92f3d81a05216798a7337
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>