imagine.sketch: remove white solid background
Most of the assets, exluding some handles and indicators, had a solid
white background (for the entire image). In general, nothing should
have a solid background, but the areas around button corners, slider
track and so on should be transparent. This patch simply selected all
exportable assets and unchecked "Background color".

Change-Id: Idebc0811290e8617f1a8cf5c3ebef665f02b8b62
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>