Doc: let TextEdit grow vertically in ensureVisible() snippet
Binding its height to the height of the flickable means that it's only
as tall as the flickable, and hence cannot be interacted with beyond a
certain point. The flickable was using the correct height
(paintedHeight), which meant that the entire text could be scrolled,
just not interacted with.

Let it assume its implicitHeight so that it can grow freely.

Task-number: QTBUG-48174
Change-Id: I853d0f505466e26e3100e79ff70210fcfa247f0f
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>