Fix bug in qmltest/events/tst_touch
The "internal" MPTA was a child of the "touchArea" MPTA and was
covering the region (50,0,100x50). Thus, it extended vertically from
the top down to y-pixel position 50. (QRects are inclusive)
Sending a touch event to (50,50) would therefore hit the internal MPTA,
but the test expected it to hit the outer MPTA.

Change-Id: I648f004c56b59f4968d54bf14d2c1908610e93ed
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>