Fix the QML context of popup background dimmers
We were accidentally passing QQuickOverlay as a QML context parent for
dynamically created popup background dimmers. There is only one overlay
per window, and the overlay is not destroyed when a popup is destroyed.
QQuickOverlay is just a visual parent item for popup background dimmers,
but the actual owner is QQuickPopup. Make sure to pass the popup instance
as a QML context parent to ensure that the popup background dimmer's QML
context is cleaned up when the popup is destroyed.

Task-number: QTBUG-63672
Change-Id: I5cabbcef5035cef40af3c9edbcd8f7ae3157bd8a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>