Refactor Call/Construct instructions
Give them a pointer to argc and argv instead of a pointer
to a full callData. Like this we can construct the callData
at the end of the JS stack and avoid the need to create an
additional copy in VME::exec().

This also opens up the option of completely avoiding all copies for
calls into runtime methods.

Also make sure that the calldata we pass into other functions is
always at the top of the JS stack.

Change-Id: I3d0eb49f7bfd7adb9ddabb213422087c66e5a520
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>