QQuickIconImage: fix crash with image providers
QQuickImageBase::load() cannot call until the component creation is
completed, because it requires an associated QML engine. All calls to
updateIcon() are guarded with the appropriate isComponentComplete()
check, but the sourceSizeChanged() signal was connected directly to
updateIcon(). Establish the signal-slot connection at component
completion to avoid calling updateIcon() to soon.

Task-number: QTBUG-63959
Change-Id: I3c935291796dbae031d2728e1d51c55596a51cd0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>