QQuickButton: promote autoRepeat back to QQuickAbstractButton
The original reason why autoRepeat was demoted to QQuickButton was that
some subclasses, such as QQuickCheckBox, did not need this property.
The same was done with the checkable property. It was first demoted,
and then promoted back to QQuickAbstractButton. These can be useful for
building custom AbstractButton-based controls. Now that we want to make
auto-repeat configurable (autoRepeatInterval and autoRepeatDelay), and
the logic has to be in the base class anyway, it makes more sense to
expose all these in the base class instead of doing the property hiding
tricks for all three properties.

[ChangeLog][Controls][AbstractButton] The autoRepeat property was
promoted from Button to AbstractButton.

Change-Id: Ife105da00d7d87a74fc8160ec35762b6f71d3ad9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>