Add Control::horizontal|verticalPadding
A common pattern for paddings:

    padding: X
    topPadding: Y
    bottomPadding: Y

This pattern can be simplified by providing separate horizontalPadding
and verticalPadding properties:

    padding: X
    horizontalPadding: Y

The actual style changes will be applied separately. As a bonus, the
size of QQuickControlPrivate can be reduced a bit since we can move the
(has)top/left/right/bottomPadding members to ExtraData.

[ChangeLog][Controls][Control] Added horizontalPadding and verticalPadding
properties as a convenient way to set both left and right, or top and bottom
paddings in one go.

Change-Id: I3a93b6f9ab988d806b1f263e74b6bd6b5427fbbe
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>