Default: cleanup an unnecessary extra item from the background
The size (and position) of the background item is respected if
explicitly defined, so the wrapper item is not needed. This does
not reduce or simplify bindings, so the performance boost is not
remarkable, but this can improve the qmlbench results by a frame
or two depending on the hardware. ;)

Change-Id: Ibd45cf02a3eb30b48b4af86e369a1db4cc5d3864
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>