Optimize QQuickTumbler
Listen to the relevant view changes once in QQuickTumbler and cache
the offset/contentY instead of every QQuickTumblerAttached instance
doing its own geometry and child tracking, and cache the view offset
or contentY so that they are readily available while calculating
displacements for attached property objects.

This gives a 5% boost (40->42 frames in qmlbench) on TX1 in release
mode.

Change-Id: If1a77468e812e65bc07f32216ff9bf2e1dc5b935
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>