Default: merge CheckIndicator back to CheckBox & CheckDelegate
The internal CheckIndicator helper was introduced together with
CheckDelegate in 1acb34a, because we naturally wanted to share the
indicator instead of duplicating it. This change is controversial,
because it leads to duplicate code, but keeping the indicator
definitions inline is clearly faster. This is not seen as a huge
problem for the Default style, because the indicator is not too
complicated. Basically, this fixes a ~5% performance regression
introduced by 1acb34a.

Before:

    running: qmlbench/benchmarks/auto/creation/quick.controls2/delegates_checkbox.qml
        [...]
        Average: 72.8  frames; using  samples; MedianAll=73; StdDev=1.48324, CoV=0.0203742

After:

    running: qmlbench/benchmarks/auto/creation/quick.controls2/delegates_checkbox.qml
        [...]
        Average: 77  frames; using  samples; MedianAll=77; StdDev=1.41421, CoV=0.0183664

Change-Id: Ibee0e29e83a64ee4a6a772a90b1784a9c8c715bb
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>