Default: micro-optimize MenuItem
A binding to "<id>.<property>" is a tiny bit faster than a binding to
"parent.<property>". ID is a direct reference, like a pointer in C++,
whereas "parent" is a property so the binding is effectively depending
on two properties: "<property>.<property>".

The gain is not huge, but enough to show up in qmlbench. ;)

Before:
    running: qmlbench/benchmarks/auto/creation/quick.controls2/delegates_menuitem.qml
        [...] (10 repeats, all results between 65-66)
        Average: 65.5  frames; using  samples; MedianAll=66; StdDev=0.527046, CoV=0.00804651

After:
    running: qmlbench/benchmarks/auto/creation/quick.controls2/delegates_menuitem.qml
        [...] (10 repeats, all results between 66-67)
        Average: 66.6  frames; using  samples; MedianAll=67; StdDev=0.516398, CoV=0.00775372

Change-Id: Ib161fad3b11e19a4bb0f39900639131f36838d2a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>