ListView: don't stop moving the highlight because of model updates
In QQuickItemViewPrivate::applyModelChanges(), if
moveReason = QQuickItemViewPrivate::Other,
then QQuickItemView::trackedPositionChanged() will fail to call
d->setPosition(pos), which is normally what keeps the Flickable
moving for a while.  Leave the reason as-is (it will be SetIndex in
this case), so as not to forget that we were actually trying to move
down.  Updating the model was just a side-effect of that: either
because some QML code was trying to append to the model or because
fetchMore() was called.

Task-number: QTBUG-61269
Task-number: QTBUG-62864
Change-Id: I3fd402469950d6c12e6a8d6e42be83ea4f54776a
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>