XMLHttpRequest: Implement XHR.overrideMimeType()
It was missing from QQmlXMLHttpRequestCtor, making life painful for
XHR users.

It's a method that has existed in every XHR specification, and is
required, if the goal is to be compliant with any XHR web standard.

[ChangeLog][Qml][XMLHttpRequest] Added missing overrideMimeType(mime)
method. This function can be used to force the XHR object to parse the
data in HTTP responses differently than what the server suggests.

Done-with: Oliver Eftevaag <oliver.eftevaag@qt.io>
Task-number: QTBUG-53709
Change-Id: I9f8ff37e1604b95306a85fc7e64db6d111b9e069
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>