Never create pointer events from mouseGrabberItem()
Normally, this was not a problem, but it is problematic during
QQuickWindow destruction:
The list of pointer event instances are destroyed, but later the grabber
is removed (through call to removeGrabber()).  This queries the
mouseGrabberItem(), which would create a new pointer event instance.

It also has the benefit that d->pointerEventInstances are now only
populated due to actual incoming events.

Task-number: QTBUG-61434
Change-Id: I4e7b6f5643f3b971138a1f7c7237ee734d29783c
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>