DialogButtonBox: workaround implicit size calculation with one button
When there's only one button in the dialog button box, the Default and
Universal styles resize the button to cover half of the button box.
This works in typical scenarios when the dialog button box is assigned
as a footer of a dialog, and thus, gets resized together with the
dialog. However, if the dialog button box is placed into a layout, or
otherwise not given an explicit size, the implicit size calculation
loops until it reaches zero.

 1) button box gets the implicit size of the content (one button)
 2) button box resizes the button to cover half of the box width
 3) button box re-calculates its implicit size => step 1

Avoid the problem by providing a reasonable hard-coded implicit size
for this special case.

Notice that this is just a temporary workaround to avoid the problem.
This can be fixed properly in dev by providing separate contentWidth
and contentHeight properties that cleanly propagate the content size
to QML.

Task-number: QTBUG-59719
Change-Id: I552e0824ae6bff26b570c699252a3e4f09bd3397
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>